chore(English.toml): add new tools translations and fix typo in key names
The commit adds new translations for tools in the English.toml file. It includes translations for "Teleport to marker", "Jetpack", "Disable ENEX markers", "Teleport to actor", "Teleport to vehicle", "Teleport to object", and "Antifreez". Additionally, it fixes a typo in the key names for the gizmo shortcuts, changing "Guizmo" to "Gizmo".
Fix issue: #26