feat(FilesystemLuaWrapper.cpp): add support for removing directories and files within the LDYOM_PATH directory
The `filesystemWrapper` function in `FilesystemLuaWrapper.cpp` has been updated to include a new function called `remove`. This function takes a Lua path as an argument and removes the corresponding directory or file within the `LDYOM_PATH` directory. Before removing the directory or file, the function checks if the path is a subpath of `LDYOM_PATH`. If it is not, the function returns -1. Otherwise, it removes the directory or file using the `remove_all` function from the `std::filesystem` library. This change allows for easier management of directories and files within the `LDYOM_PATH` directory.