feat: add ImGui::RenamePopup with char* and std::string overloads
This commit introduces a new `RenamePopup` widget for ImGui, allowing
in-place renaming with a text input field inside a popup. It includes
overloads for both `char*` and `std::string` buffers. The popup closes
on Enter key press or clicking outside the input field. Also includes
`<imgui_stdlib.h>` to be able to use `std::string` with ImGui InputText.