feat: add objective execution and ktcoro_wait dependency
This commit introduces the ability to execute objectives and adds
`ktcoro_wait.hpp` as a dependency. The `Objective` class now includes
an `execute` method that calls the `callExecutor` method of its
content. A default executor callback has been added to the
`Objective` constructor. The `wait_signal` objective has been updated
to include an executor callback.